# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Web Umkm",
        "description": "This For API System,And You Can Try It",
        "contact": {
            "email": "muhammadalghifari4321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/barang": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Barang",
                "description": "Looking Data Barang",
                "operationId": "getBarang",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data Barang Admin",
                "description": "Insert Data Barang Admin",
                "operationId": "insertBarang",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_barang": {},
                                    "harga_barang": {},
                                    "toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/barang/{id}": {
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data Barang Admin",
                "description": "Update Data Barang Admin",
                "operationId": "updateBarang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_barang": {},
                                    "harga_barang": {},
                                    "toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data Barang Admin",
                "description": "Delete Data Barang Admin",
                "operationId": "deleteBarang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Login Admin,for admin i guess",
                "description": "Login Admin,for admin i guess",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {},
                                    "password": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Register Admin,for admin i guess",
                "description": "Register Admin,for admin i guess",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {},
                                    "email": {},
                                    "password": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/toko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Toko",
                "description": "Looking Data Toko",
                "operationId": "getToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data Toko Admin",
                "description": "Insert Data Toko Admin",
                "operationId": "insertToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_toko": {},
                                    "daerah": {},
                                    "pemilik": {},
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/toko/{id}": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get By Id Data Toko Admin",
                "description": "Get By Id Data Toko Admin",
                "operationId": "getByIdToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Get Data And Relation",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data Toko Admin",
                "description": "Update Data Toko Admin",
                "operationId": "updateToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_toko": {},
                                    "daerah": {},
                                    "pemilik": {},
                                    "jenis_toko": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data Toko Admin",
                "description": "Delete Data Toko Admin",
                "operationId": "deleteToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Account Admin",
                "description": "Looking For Account Admin Login",
                "operationId": "users",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Welcome page"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Projects UMKM",
            "description": "Endpoint API Project"
        }
    ],
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}